requirements.txt
groq
sentence-transformers
faiss-cpu
gradio
pypdf
python-dotenv
numpy
